Rename before rebasing by other branch
