+++ convert/src/label_studio_to_badgerdoc/labelstudio_format/__init__.py
@@ -1,1 +1,1 @@
-from .labelstudio_format import LabelStudioFormat
+from .label_studio_format import LabelStudioFormat
