+++ convert/tests/test_label_studio/test_export.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from src.label_studio_to_badgerdoc.models.label_studio_models import (
     LabelStudioModel,
 )
-from src.label_studio_to_badgerdoc.labelstudio_format.labelstudio_format import LabelStudioFormat
+from src.label_studio_to_badgerdoc.labelstudio_format.label_studio_format import LabelStudioFormat
 from src.label_studio_to_badgerdoc.badgerdoc_format import annotation_converter_practic
 from src.label_studio_to_badgerdoc.models import bd_annotation_model_practic
 from src.label_studio_to_badgerdoc.models.bd_tokens_model import Page
